requirements.txt
python-dotenv>=0.21.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pandas>=2.0.0
pytest>=7.0.0
httpx>=0.24.0